Dynamic Address Allocation
This branch adds the ability to dynamically assign address to the source code without
specifying the destination address (numerical alias name). This is achieved by changing
the behavior of "s" and "LABLE" alias and with the addition of Jmp instructions which
use stack head as jump destination. This also changes the order in which the source cfg
files should be executed, main should always be executed first followed by the
dependencies and function libraries. In order to avoid this issue "origin" alias should be
used in the cfg to specify which position the code following it should go.