Load and Store Using Stack
This commit adds "store_vfs" and "load_vfs" which use the items on the stack to read and write to the database.
store_vfs needs two items on the stack, one item is the key name and the other is the value associated to the key.
load_vfs on the other hand needs one item, on the stack which is the key, and load it's associated value into stack.